+++ src/main/java/VisionApi/JSONClasses/LatLng.java
@@ -4,6 +4,6 @@ package VisionApi.JSONClasses;
  * Created by sierra on 4/22/16.
  */
 public class LatLng {
-    public int latitude;
-    public int longitude;
+    public double latitude;
+    public double longitude;
 }
